Refactor rendering and filter systems
