+++ src/filters/defaults/blur/BlurFilter.ts
@@ -4,7 +4,7 @@ import { deprecation, v8_0_0 } from '../../../utils/logging/deprecation';
 import { Filter } from '../../Filter';
 import { BlurFilterPass } from './BlurFilterPass';
 
-import type { RenderSurface } from '../../../rendering/renderers/gpu/renderTarget/GpuRenderTargetSystem';
+import type { RenderSurface } from '../../../rendering/renderers/shared/renderTarget/RenderTargetSystem';
 import type { BLEND_MODES } from '../../../rendering/renderers/shared/state/const';
 import type { Texture } from '../../../rendering/renderers/shared/texture/Texture';
 import type { FilterOptions } from '../../Filter';
