+++ src/rendering/renderers/gl/GlBackBufferSystem.ts
@@ -5,8 +5,7 @@ import { TextureSource } from '../shared/texture/sources/TextureSource';
 import { Texture } from '../shared/texture/Texture';
 import { GlProgram } from './shader/GlProgram';
 
-import type { RgbaArray } from '../../../color/Color';
-import type { RenderSurface } from '../gpu/renderTarget/GpuRenderTargetSystem';
+import type { RenderOptions } from '../shared/system/AbstractRenderer';
 import type { System } from '../shared/system/System';
 import type { WebGLRenderer } from './WebGLRenderer';
 
@@ -65,7 +64,7 @@ export class GlBackBufferSystem implements System
 
     private _backBufferTexture: Texture;
     private readonly _renderer: WebGLRenderer;
-    private _targetTexture: Texture;
+    private _targetTexture: TextureSource;
     private _useBackBufferThisRender = false;
     private _antialias: boolean;
 
@@ -82,22 +81,23 @@ export class GlBackBufferSystem implements System
         this._antialias = antialias;
     }
 
-    protected renderStart({ target, clear, clearColor }: { target: RenderSurface, clear: boolean, clearColor: RgbaArray })
+    /**
+     * This is called before the RenderTargetSystem is started. This is where
+     * we replace the target with the back buffer if required.
+     * @param options - The options for this render.
+     */
+    protected renderStart(options: RenderOptions)
     {
-        this._useBackBufferThisRender = this.useBackBuffer && !!target;
+        this._useBackBufferThisRender = this.useBackBuffer && !!options.target;
 
-        if (this.useBackBuffer)
+        if (this._useBackBufferThisRender)
         {
-            const renderTarget = this._renderer.renderTarget.getRenderTarget(target);
+            const renderTarget = this._renderer.renderTarget.getRenderTarget(options.target);
 
             this._targetTexture = renderTarget.colorTexture;
 
-            target = this._getBackBufferTexture(renderTarget.colorTexture);
+            options.target = this._getBackBufferTexture(renderTarget.colorTexture);
         }
-
-        clearColor ??= this._renderer.background.colorRgba;
-
-        this._renderer.renderTarget.start(target, clear, clearColor);
     }
 
     protected renderEnd()
@@ -125,24 +125,22 @@ export class GlBackBufferSystem implements System
         gl.drawArrays(gl.TRIANGLES, 0, 3);
     }
 
-    private _getBackBufferTexture(targetTexture: Texture)
+    private _getBackBufferTexture(targetSourceTexture: TextureSource)
     {
-        const source = targetTexture.source;
-
         this._backBufferTexture = this._backBufferTexture || new Texture({
             source: new TextureSource({
-                width: source.width,
-                height: source.height,
-                resolution: source._resolution,
+                width: targetSourceTexture.width,
+                height: targetSourceTexture.height,
+                resolution: targetSourceTexture._resolution,
                 antialias: this._antialias,
             }),
         });
 
         // this will not resize if its the same size already! No extra check required
         this._backBufferTexture.source.resize(
-            source.width,
-            source.height,
-            source._resolution,
+            targetSourceTexture.width,
+            targetSourceTexture.height,
+            targetSourceTexture._resolution,
         );
 
         return this._backBufferTexture;
