+++ src/rendering/renderers/gl/GlRenderTarget.ts
@@ -2,12 +2,9 @@ export class GlRenderTarget
 {
     public width = -1;
     public height = -1;
-
+    public msaa = false;
     public framebuffer: WebGLFramebuffer;
     public resolveTargetFramebuffer: WebGLFramebuffer;
     public msaaRenderBuffer: WebGLRenderbuffer[] = [];
     public depthStencilRenderBuffer: WebGLRenderbuffer;
-
-    public msaa = false;
-    public dirtyId = -1;
 }
