+++ src/scene/container/LayerPipe.ts
@@ -36,10 +36,8 @@ export class LayerPipe implements InstructionPipe<LayerGroup>
         if (!layerGroup.isRenderable) return;
 
         this._renderer.globalUniforms.push({
-            projectionData: this._renderer.renderTarget.renderTarget,
             worldTransformMatrix: layerGroup.worldTransform,
             worldColor: layerGroup.worldColor,
-
         });
 
         executeInstructions(layerGroup, this._renderer.renderPipes);
